move contract_transact to Ethereum module
